Make each page scrollable
This fixes an issue where when moving to the next page, the page would
be shown with the same scroll position as the previous page. This was
because all pages used the same scroll container.

This makes it so that each page is in its own scroll container. So when
you transition to the next page it will be at the top.