Preallocate images
This changes it so that images are preallocated rather than just keeping
the URLs. This means that we don't have to wait on an image to load each
time we cycle through pages. However we do still recycle the use of the
canvases.

For now we are not preallocating *all* images. This is because some
books can be large (500+ pages perhaps) and I'm worried about using up
that much memory.