Prevent accidental clicks on desktop
I want to test this thoroughly on mobile before merging it. Closes #41

Instead of having a `click` event which seems to always fire on desktop
after a pointer event, this instead simulates what that means with a
`tap` event if there has been no panning.

It's unclear to me if a little bit of panning occurs on mobile all the
time, in which case we'll need to fine-tune this a little bit. So going
to leave as an open PR until it can be tested.